+++ packages/api/internal/db/schema/tier.go
@@ -13,9 +13,9 @@ type Tier struct {
 func (Tier) Fields() []ent.Field {
 	return []ent.Field{
 		field.Text("id").Immutable().Unique(),
-		field.Int8("vcpu"),
-		field.Int8("ram_mb"),
-		field.Int8("disk_mb"),
+		field.Int64("vcpu"),
+		field.Int64("ram_mb"),
+		field.Int64("disk_mb"),
 	}
 }
